+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/CDDL.hs
@@ -96,8 +96,8 @@ header_body :: Rule
 header_body =
   "header_body"
     =:= arr
-      [ "block_number" ==> block_no
-      , "slot" ==> slot_no
+      [ "block_number" ==> block_number
+      , "slot" ==> slot
       , "prev_hash" ==> (hash32 / VNil)
       , "issuer_vkey" ==> vkey
       , "vrf_vkey" ==> vrf_vkey
@@ -118,11 +118,11 @@ transaction_body =
       [ idx 0 ==> maybe_tagged_set transaction_input
       , idx 1 ==> arr [0 <+ a transaction_output]
       , idx 2 ==> coin
-      , opt (idx 3 ==> slot_no)
+      , opt (idx 3 ==> slot)
       , opt (idx 4 ==> certificates)
       , opt (idx 5 ==> withdrawals)
       , opt (idx 7 ==> auxiliary_data_hash)
-      , opt (idx 8 ==> slot_no) -- Validity interval start
+      , opt (idx 8 ==> slot) -- Validity interval start
       , opt (idx 9 ==> mint)
       , opt (idx 11 ==> script_data_hash)
       , opt (idx 13 ==> maybe_tagged_nonempty_set transaction_input)